requirements.txt
feedparser==6.0.11
PyYAML==6.0.1
requests==2.31.0
beautifulsoup4==4.12.3
